fix: mypy ratchet (4k→0) + UX audit round 3 (10 fixes) (#1272)
* chore(backend): mypy ratchet — 4,000 → 0 errors across 1,289 source files

Systematically eliminate all mypy type-checking errors without changing
runtime logic. Fixes include targeted `# type: ignore[error-code]`
annotations, proper type narrowing with assertions, numpy scalar casts,
Pydantic v2 Field parameter updates, and union type annotations.


* fix(frontend): UX audit round 3 — 10 findings addressed

Fixes from Antigravity (Gemini 2.5 Pro) UX audit round 3:

- /schedule: person name now links to /my-schedule?person=id
- /logout: new page that calls logout() and redirects to /login
- /compliance: remove double-scaling on coverage rate (6363% → 63.6%)
- /activities: show actual API error instead of generic "already exists"
- /people Manage tab: row click now opens edit modal (onRowClick prop)
- /admin/people: row click opens edit modal (was only toggling checkbox)
- /admin/scheduling: constraint counter reads API data, not empty config
- Mobile nav: add missing "Rotations" link
- Desktop nav: widen max-width to reduce clipping at 1440px
- Tooltip: div→span to fix hydration mismatch in <p> contexts


* fix(frontend): address Codex P2 feedback on PR #1272

- ActivityFormModal: normalize API detail to string before rendering
  (FastAPI can return array/object for validation errors)
- logout page: use .finally() so redirect always fires even on rejection


---------

+++ backend/app/scheduling/periodicity/anti_churn.py
@@ -365,8 +365,8 @@ def time_crystal_objective(
 
     if len(churn_by_person) > 0:
         churn_values = list(churn_by_person.values())
-        mean_churn = np.mean(churn_values)
-        std_churn = np.std(churn_values)
+        mean_churn = float(np.mean(churn_values))
+        std_churn = float(np.std(churn_values))
 
         # Fairness = 1 - (coefficient_of_variation)
         # CV = std / mean, measures relative variability
